+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   * Either from QGIS when "save as" is used
   * Or when the QGIS project has the `lizmap_user` project variable but no CFG file
 * Some Python cleaning about handling file path, related to CFG file for instance
+* When running Lizmap Web Client 3.6, the QGIS server must be OK before going in other tabs of the plugin
 
 ## 3.11.1 - 2023-03-21